add separate compatStore
