+++ pkg/config/BUILD
@@ -6,6 +6,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "apiserver.go",
+        "compatStore.go",
         "fsstore.go",
         "kind.go",
         "manager.go",
